Refactor parseBool and parseMap into tuil/strings.go
